pyproject.toml
[tool.poetry]
name = "fastapi-example"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = "^0.18.2"
typer = "^0.6.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
mypy = "^0.942"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
